Base class for listX in Video
